move addType to it's own hook useCreateType, implement it and remove unnecessary eventId
